[1.1.0] Fix warning

+++ report.c
@@ -371,7 +371,7 @@ report_columns_set_str (const char *i_str, const char *s_str, const char *hdr)
 	size_t min_len = 1 + strlen(hdr);
 	ok = ok && 1 == sscanf (i_str, "%d", &int1);
 	ok = ok && 1 == sscanf (s_str, "%d", &int2);
-	ok = ok && report_columns_set (int1, (int2 < min_len? min_len: int2), hdr);
+	ok = ok && report_columns_set (int1, ((size_t)int2 < min_len? (int)min_len: int2), hdr);
 	return ok;
 }
